♻️ refactor: hoist Route resolver to shared RouteResolver
Extracts HomeView.routeDestination + ScenarioEditorHost into a
tab-agnostic RouteResolver view so all four tab NavigationStacks can
share one Route universe (ADR-016 D3) — a .scenarioDetail / .simulation
/ .results(detail) push from the Search or History tab resolves
identically to one from Home. Behavior-preserving; HomeView's single
stack now resolves via RouteResolver(route:).
